clean: alarm status views/url
