--- raspberrypi_central/core/app/camera/views/home.py
@@ -1,18 +1,0 @@
-from typing import List
-
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.views.generic import TemplateView
-
-from alarm.models import AlarmStatus
-
-
-class AlarmHome(LoginRequiredMixin, TemplateView):
-    template_name = "alarm/home.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        alarms: List[AlarmStatus] = AlarmStatus.objects.select_related('device__location').order_by('device__device_id')
-        context['alarms'] = alarms
-
-        return context
